Move traits and impls into submodules
