+++ src/lib.rs
@@ -160,333 +160,15 @@
 #[cfg(feature="use_complex")]
 extern crate num_complex;
 
-#[cfg(feature="no_std")]
-use core as std;
-#[cfg(feature="no_std")]
-#[cfg_attr(feature="no_std", allow(unused_imports))] // HACK: seems to be a bug in this lint!
-use core::num::Float;
-#[cfg(feature="use_complex")]
-use num_complex::Complex;
+mod abs_diff_eq;
+mod relative_eq;
+mod ulps_eq;
 
 mod macros;
 
-/// Equality that is defined using the absolute difference of two numbers.
-pub trait AbsDiffEq: PartialEq {
-    /// Used for specifying relative comparisons.
-    type Epsilon;
-
-    /// The default tolerance to use when testing values that are close together.
-    ///
-    /// This is used when no `epsilon` value is supplied to the `abs_diff_eq!`, `relative_eq!`, or
-    /// `ulps_eq!` macros.
-    fn default_epsilon() -> Self::Epsilon;
-
-    /// A test for equality that uses the absolute difference to compute the approximate
-    /// equality of two numbers.
-    fn abs_diff_eq(&self, other: &Self, epsilon: Self::Epsilon) -> bool;
-
-    /// The inverse of `ApproxEq::abs_diff_eq`.
-    fn abs_diff_ne(&self, other: &Self, epsilon: Self::Epsilon) -> bool {
-        !Self::abs_diff_eq(self, other, epsilon)
-    }
-}
-
-impl<'a, T: AbsDiffEq> AbsDiffEq for &'a T {
-    type Epsilon = T::Epsilon;
-
-    #[inline]
-    fn default_epsilon() -> Self::Epsilon {
-        T::default_epsilon()
-    }
-
-    #[inline]
-    fn abs_diff_eq(&self, other: &&'a T, epsilon: T::Epsilon) -> bool {
-        T::abs_diff_eq(*self, *other, epsilon)
-    }
-}
-
-impl<'a, T: AbsDiffEq> AbsDiffEq for &'a mut T {
-    type Epsilon = T::Epsilon;
-
-    #[inline]
-    fn default_epsilon() -> Self::Epsilon {
-        T::default_epsilon()
-    }
-
-    #[inline]
-    fn abs_diff_eq(&self, other: &&'a mut T, epsilon: T::Epsilon) -> bool {
-        T::abs_diff_eq(*self, *other, epsilon)
-    }
-}
-
-macro_rules! impl_unsigned_abs_diff_eq {
-    ($T:ident, $default_epsilon:expr) => {
-        impl AbsDiffEq for $T {
-            type Epsilon = $T;
-
-            #[inline]
-            fn default_epsilon() -> $T { $default_epsilon }
-
-            #[inline]
-            fn abs_diff_eq(&self, other: &$T, epsilon: $T) -> bool {
-                (if self > other { self - other } else { other - self }) <= epsilon
-            }
-        }
-    }
-}
-
-impl_unsigned_abs_diff_eq!(u8, 0);
-impl_unsigned_abs_diff_eq!(u16, 0);
-impl_unsigned_abs_diff_eq!(u32, 0);
-impl_unsigned_abs_diff_eq!(u64, 0);
-impl_unsigned_abs_diff_eq!(usize, 0);
-
-macro_rules! impl_signed_abs_diff_eq {
-    ($T:ident, $default_epsilon:expr) => {
-        impl AbsDiffEq for $T {
-            type Epsilon = $T;
-
-            #[inline]
-            fn default_epsilon() -> $T { $default_epsilon }
-
-            #[inline]
-            fn abs_diff_eq(&self, other: &$T, epsilon: $T) -> bool {
-                $T::abs(self - other) <= epsilon
-            }
-        }
-    }
-}
-
-impl_signed_abs_diff_eq!(i8, 0);
-impl_signed_abs_diff_eq!(i16, 0);
-impl_signed_abs_diff_eq!(i32, 0);
-impl_signed_abs_diff_eq!(i64, 0);
-impl_signed_abs_diff_eq!(isize, 0);
-impl_signed_abs_diff_eq!(f32, std::f32::EPSILON);
-impl_signed_abs_diff_eq!(f64, std::f64::EPSILON);
-
-#[cfg(feature="use_complex")]
-impl<T: AbsDiffEq> AbsDiffEq for Complex<T>
-    where T::Epsilon: Clone
-{
-    type Epsilon = T::Epsilon;
-
-    #[inline]
-    fn default_epsilon() -> T::Epsilon {
-        T::default_epsilon()
-    }
-
-    #[inline]
-    fn abs_diff_eq(&self, other: &Complex<T>, epsilon: T::Epsilon) -> bool {
-        T::abs_diff_eq(&self.re, &other.re, epsilon.clone()) &&
-        T::abs_diff_eq(&self.im, &other.im, epsilon.clone())
-    }
-}
-
-/// Equality comparisons between two numbers using both the absolute difference and
-/// relative based comparisons.
-pub trait RelativeEq: AbsDiffEq {
-    /// The default relative tolerance for testing values that are far-apart.
-    ///
-    /// This is used when no `max_relative` value is supplied to the `relative_eq` macro.
-    fn default_max_relative() -> Self::Epsilon;
-
-    /// A test for equality that uses a relative comparison if the values are far apart.
-    fn relative_eq(&self,
-                   other: &Self,
-                   epsilon: Self::Epsilon,
-                   max_relative: Self::Epsilon)
-                   -> bool;
-
-    /// The inverse of `ApproxEq::relative_eq`.
-    fn relative_ne(&self,
-                   other: &Self,
-                   epsilon: Self::Epsilon,
-                   max_relative: Self::Epsilon)
-                   -> bool {
-        !Self::relative_eq(self, other, epsilon, max_relative)
-    }
-}
-
-impl<'a, T: RelativeEq> RelativeEq for &'a T {
-    #[inline]
-    fn default_max_relative() -> Self::Epsilon {
-        T::default_max_relative()
-    }
-
-    #[inline]
-    fn relative_eq(&self, other: &&'a T, epsilon: T::Epsilon, max_relative: T::Epsilon) -> bool {
-        T::relative_eq(*self, *other, epsilon, max_relative)
-    }
-}
-
-impl<'a, T: RelativeEq> RelativeEq for &'a mut T {
-    #[inline]
-    fn default_max_relative() -> Self::Epsilon {
-        T::default_max_relative()
-    }
-
-    #[inline]
-    fn relative_eq(&self,
-                   other: &&'a mut T,
-                   epsilon: T::Epsilon,
-                   max_relative: T::Epsilon)
-                   -> bool {
-        T::relative_eq(*self, *other, epsilon, max_relative)
-    }
-}
-
-// Implementation based on: [Comparing Floating Point Numbers, 2012 Edition]
-// (https://randomascii.wordpress.com/2012/02/25/comparing-floating-point-numbers-2012-edition/)
-macro_rules! impl_relative_eq {
-    ($T:ident, $U:ident) => {
-        impl RelativeEq for $T {
-            #[inline]
-            fn default_max_relative() -> $T { std::$T::EPSILON }
-
-            #[inline]
-            fn relative_eq(&self, other: &$T, epsilon: $T, max_relative: $T) -> bool {
-                // Handle same infinities
-                if self == other {
-                    return true;
-                }
-
-                let abs_diff = $T::abs(self - other);
-
-                // For when the numbers are really close together
-                if abs_diff <= epsilon {
-                    return true;
-                }
-
-                let abs_self = $T::abs(*self);
-                let abs_other = $T::abs(*other);
-
-                // Handle oppsite infinities
-                if abs_self == abs_other && abs_diff == abs_self {
-                    return false;
-                }
-
-                let largest = if abs_other > abs_self { abs_other } else { abs_self };
-
-                // Use a relative difference comparison
-                abs_diff <= largest * max_relative
-            }
-        }
-    }
-}
-
-impl_relative_eq!(f32, i32);
-impl_relative_eq!(f64, i64);
-
-#[cfg(feature="use_complex")]
-impl<T: RelativeEq> RelativeEq for Complex<T>
-    where T::Epsilon: Clone
-{
-    #[inline]
-    fn default_max_relative() -> T::Epsilon {
-        T::default_max_relative()
-    }
-
-    #[inline]
-    fn relative_eq(&self,
-                   other: &Complex<T>,
-                   epsilon: T::Epsilon,
-                   max_relative: T::Epsilon)
-                   -> bool {
-        T::relative_eq(&self.re, &other.re, epsilon.clone(), max_relative.clone()) &&
-        T::relative_eq(&self.im, &other.im, epsilon.clone(), max_relative.clone())
-    }
-}
-
-/// Equality comparisons between two numbers using both the absolute difference and ULPs
-/// (Units in Last Place) based comparisons.
-pub trait UlpsEq: AbsDiffEq {
-    /// The default ULPs to tolerate when testing values that are far-apart.
-    ///
-    /// This is used when no `max_ulps` value is supplied to the `ulps_eq` macro.
-    fn default_max_ulps() -> u32;
-
-    /// A test for equality that uses units in the last place (ULP) if the values are far apart.
-    fn ulps_eq(&self, other: &Self, epsilon: Self::Epsilon, max_ulps: u32) -> bool;
-
-    /// The inverse of `ApproxEq::ulps_eq`.
-    fn ulps_ne(&self, other: &Self, epsilon: Self::Epsilon, max_ulps: u32) -> bool {
-        !Self::ulps_eq(self, other, epsilon, max_ulps)
-    }
-}
-
-impl<'a, T: UlpsEq> UlpsEq for &'a T {
-    #[inline]
-    fn default_max_ulps() -> u32 {
-        T::default_max_ulps()
-    }
-
-    #[inline]
-    fn ulps_eq(&self, other: &&'a T, epsilon: T::Epsilon, max_ulps: u32) -> bool {
-        T::ulps_eq(*self, *other, epsilon, max_ulps)
-    }
-}
-
-impl<'a, T: UlpsEq> UlpsEq for &'a mut T {
-    #[inline]
-    fn default_max_ulps() -> u32 {
-        T::default_max_ulps()
-    }
-
-    #[inline]
-    fn ulps_eq(&self, other: &&'a mut T, epsilon: T::Epsilon, max_ulps: u32) -> bool {
-        T::ulps_eq(*self, *other, epsilon, max_ulps)
-    }
-}
-
-// Implementation based on: [Comparing Floating Point Numbers, 2012 Edition]
-// (https://randomascii.wordpress.com/2012/02/25/comparing-floating-point-numbers-2012-edition/)
-macro_rules! impl_ulps_eq {
-    ($T:ident, $U:ident) => {
-        impl UlpsEq for $T {
-            #[inline]
-            fn default_max_ulps() -> u32 { 4 }
-
-            #[inline]
-            fn ulps_eq(&self, other: &$T, epsilon: $T, max_ulps: u32) -> bool {
-                // For when the numbers are really close together
-                if $T::abs_diff_eq(self, other, epsilon) {
-                    return true;
-                }
-
-                // Trivial negative sign check
-                if self.signum() != other.signum() {
-                    return false;
-                }
-
-                // ULPS difference comparison
-                let int_self: $U = unsafe { std::mem::transmute(*self) };
-                let int_other: $U = unsafe { std::mem::transmute(*other) };
-
-                $U::abs(int_self - int_other) < max_ulps as $U
-            }
-        }
-    }
-}
-
-impl_ulps_eq!(f32, i32);
-impl_ulps_eq!(f64, i64);
-
-#[cfg(feature="use_complex")]
-impl<T: UlpsEq> UlpsEq for Complex<T>
-    where T::Epsilon: Clone
-{
-    #[inline]
-    fn default_max_ulps() -> u32 {
-        T::default_max_ulps()
-    }
-
-    #[inline]
-    fn ulps_eq(&self, other: &Complex<T>, epsilon: T::Epsilon, max_ulps: u32) -> bool {
-        T::ulps_eq(&self.re, &other.re, epsilon.clone(), max_ulps) &&
-        T::ulps_eq(&self.im, &other.im, epsilon.clone(), max_ulps)
-    }
-}
+pub use abs_diff_eq::AbsDiffEq;
+pub use relative_eq::RelativeEq;
+pub use ulps_eq::UlpsEq;
 
 /// The requisite parameters for testing for approximate equality using a
 /// absolute difference based comparison.
